requirements.txt
Flask
gunicorn

PyJWT==2.8.0
uuid==1.30